// frontend/web/src/services/API/proserve-wb-projects-api/models/ProjectAccount.ts
export interface ProjectAccount {
    /**
     * Unique ID of the account.
     * @type {string}
     * @memberof ProjectAccount
     */
    id?: string;
    /**
     * Unique ID of the AWS account.
     * @type {string}
     * @memberof ProjectAccount
     */
    awsAccountId: string;
    /**
     * Type of the account.
     * @type {string}
     * @memberof ProjectAccount
     */
    accountType: string;
    /**
     * Name of the account.
     * @type {string}
     * @memberof ProjectAccount
     */
    accountName?: string;
    /**
     * Description of the account.
     * @type {string}
     * @memberof ProjectAccount
     */
    accountDescription?: string;
    /**
     * Point in time where account is created.
     * @type {string}
     * @memberof ProjectAccount
     */
    createDate?: string;
    /**
     * Point in time where account is last updated.
     * @type {string}
     * @memberof ProjectAccount
     */
    lastUpdateDate?: string;
    /**
     * Account status
     * @type {string}
     * @memberof ProjectAccount
     */
    accountStatus?: string;
    /**
     * The stage account represents. Possible values dev, qa, prod
     * @type {string}
     * @memberof ProjectAccount
     */
    stage?: string;
    /**
     * The reference to the associated technology platform
     * @type {string}
     * @memberof ProjectAccount
     */
    technologyId: string;
    /**
     * Region in which workbenches will be provisioned in this account.
     * @type {string}
     * @memberof ProjectAccount
     */
    region?: string;
    /**
     * Result of the last onboarding / reonboarding
     * @type {string}
     * @memberof ProjectAccount
     */
    lastOnboardingResult?: string;
    /**
     * Error message of the last onboarding / reonboarding if it failed
     * @type {string}
     * @memberof ProjectAccount
     */
    lastOnboardingErrorMessage?: string;
}